feat: [SRM-14148]: Added SLI Calculation for request SLO (#45268)
* feat:[SRM-14148]: Added metric values in sliAnalyseResponse

* feat:[SRM-14148]: Added SLIRecord Changes for simple SLO

* feat:[SRM-14148]: Added changes for review comments and fixed tests

* feat:[SRM-14148]: fixed tests

* feat:[SRM-14148]: Added changes for request SLI onboarding graph

* feat:[SRM-14148]: Changed SLIValue feilds to long

* feat:[SRM-14148]: Removed Errors cause of int to long conversion

* feat:[SRM-14148]: Added typecast for slo health entity

* feat:[SRM-14148]: Fixed tests

+++ srm-service/modules/cv-nextgen-service/service/src/main/java/io/harness/cvng/servicelevelobjective/beans/SLIMissingDataType.java
@@ -25,22 +25,18 @@ public enum SLIMissingDataType {
     switch (this) {
       case GOOD:
         return SLIValue.builder()
-            .goodCount((int) (goodCount + missingDataCount))
-            .badCount((int) badCount)
-            .total((int) (totalMinutes - disabledMinutes))
+            .goodCount(goodCount + missingDataCount)
+            .badCount(badCount)
+            .total(totalMinutes - disabledMinutes)
             .build();
       case BAD:
         return SLIValue.builder()
-            .goodCount((int) goodCount)
-            .badCount((int) (badCount + missingDataCount))
-            .total((int) (totalMinutes - disabledMinutes))
+            .goodCount(goodCount)
+            .badCount(badCount + missingDataCount)
+            .total(totalMinutes - disabledMinutes)
             .build();
       case IGNORE:
-        return SLIValue.builder()
-            .goodCount((int) goodCount)
-            .badCount((int) badCount)
-            .total((int) (goodCount + badCount))
-            .build();
+        return SLIValue.builder().goodCount(goodCount).badCount(badCount).total(goodCount + badCount).build();
       default:
         throw new IllegalStateException("Unhanded SLIMissingDataType " + this);
     }
