+++ srm-service/modules/cv-nextgen-service/service/src/main/java/io/harness/cvng/servicelevelobjective/beans/SLIValue.java
@@ -13,9 +13,9 @@ import lombok.Value;
 @Value
 @Builder
 public class SLIValue {
-  int goodCount;
-  int badCount;
-  int total;
+  long goodCount;
+  long badCount;
+  long total;
   public double sliPercentage() {
     if (total <= 0) {
       return 100.0;
