+++ srm-service/modules/cv-nextgen-service/service/src/main/java/io/harness/cvng/servicelevelobjective/beans/SLOConsumptionBreakdown.java
@@ -25,7 +25,7 @@ public class SLOConsumptionBreakdown {
   @NotNull double weightagePercentage;
   @NotNull double sloTargetPercentage;
   @NotNull double sliStatusPercentage;
-  @NotNull int errorBudgetBurned;
+  @NotNull long errorBudgetBurned;
   @NotNull int contributedErrorBudgetBurned;
   @NotNull ProjectParams projectParams;
   String orgName;
